requirements.txt
fastapi
tortoise-orm[asyncpg]
pydantic
bcrypt
pyjwt
aerich
aiobotocore
boto3
pydantic_settings
python-dotenv
tomlkit
uvicorn